useAudio is now custom hook outside Sound.js
